LPS-131040 We don't need to have a ItemSelectorCriterionHandler for each entity, so keep only one and move it to the correct module
