--- modules/apps/journal/journal-web/src/main/java/com/liferay/journal/web/internal/item/selector/JournalArticleItemSelectorCriterionHandler.java
@@ -1,60 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.journal.web.internal.item.selector;
-
-import com.liferay.item.selector.BaseItemSelectorCriterionHandler;
-import com.liferay.item.selector.ItemSelectorCriterion;
-import com.liferay.item.selector.ItemSelectorCriterionHandler;
-import com.liferay.item.selector.ItemSelectorView;
-import com.liferay.item.selector.criteria.info.item.criterion.InfoItemItemSelectorCriterion;
-import com.liferay.item.selector.criteria.info.item.criterion.InfoItemItemSelectorCriterionHandlerUtil;
-
-import java.util.List;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Eudaldo Alonso
- */
-@Component(service = ItemSelectorCriterionHandler.class)
-public class JournalArticleItemSelectorCriterionHandler
-	extends BaseItemSelectorCriterionHandler<InfoItemItemSelectorCriterion> {
-
-	@Override
-	public Class<InfoItemItemSelectorCriterion>
-		getItemSelectorCriterionClass() {
-
-		return InfoItemItemSelectorCriterion.class;
-	}
-
-	@Override
-	public List<ItemSelectorView<InfoItemItemSelectorCriterion>>
-		getItemSelectorViews(ItemSelectorCriterion itemSelectorCriterion) {
-
-		return InfoItemItemSelectorCriterionHandlerUtil.
-			getFilteredItemSelectorViews(
-				itemSelectorCriterion,
-				super.getItemSelectorViews(itemSelectorCriterion));
-	}
-
-	@Activate
-	@Override
-	protected void activate(BundleContext bundleContext) {
-		super.activate(bundleContext);
-	}
-
-}
